Add the ajv-formats package to the image
See [1]. This allows importing the ajv-formats module in the ajv command
as follows:

   ajv -c ajv-formats <...>

Which allows validating against schemas that use the "format" [2]
keyword. The ajv-formats modules includes definitions for a bunch of
formats, such as "uri" and "date", that can be used with the "format"
keyword in schemas, for example, for string types.

[1] https://www.npmjs.com/package/ajv-formats
[2] https://json-schema.org/draft/2020-12/json-schema-validation.html#rfc.section.7