Overwrite entrypoint of base image
This prevente unexpected side effects with the entrypoint of the base
image, which is a docker-entrypoint.sh script [1], when a non-executable
argument is passed to 'docker run' (e.g. due to a typo). In these cases,
the entrypoint script would attempt to pass this argument as an argument
to the 'node' command.

Setting the entrypoint to /usr/bin/env (see [2]) ensures that 'node' is
never invoked and all arguments are attempted to be executed.

[1] https://github.com/nodejs/docker-node/blob/main/docker-entrypoint.sh
[2] https://stackoverflow.com/a/41207569/4747193